Personalised try catch errors for primarykey
Project completed.

+++ ProjetFinal_CS_BourassaMarie/DeleteProgram.cs
@@ -18,21 +18,22 @@ namespace ProjetFinal_CS_BourassaMarie
             InitializeComponent();
             FillCombo();
         }
+
         SqlConnection mydbCon;
         SqlCommand command;
         SqlDataReader reader;
 
-        private void DeleteProgram_Load(object sender, EventArgs e)
+        void connect()
         {
-
+            mydbCon = new SqlConnection(@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=
+            C:\Users\TQY_06\Documents\ProjetFinal_CS_BourassaMarie\ProjetFinal_CS_BourassaMarie\FinalProjDB.mdf;
+            Integrated Security=True");
         }
 
         void FillCombo()
         {
             comboBoxDeleteSelection.Items.Clear();
-            mydbCon = new SqlConnection(@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=
-            C:\Users\TQY_06\Documents\ProjetFinal_CS_BourassaMarie\ProjetFinal_CS_BourassaMarie\FinalProjDB.mdf;
-            Integrated Security=True");
+            connect();
             mydbCon.Open();
 
             command = new SqlCommand("SELECT * FROM Programs", mydbCon);
@@ -54,31 +55,42 @@ namespace ProjetFinal_CS_BourassaMarie
 
         private void buttonDelete_Click(object sender, EventArgs e)
         {
-            ListAllPrograms listProg = new ListAllPrograms();
-          DialogResult result =  MessageBox.Show("Do you really want to delete " + comboBoxDeleteSelection.Text + " ?", "Attention!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            mydbCon.Open();
 
-            if (result == DialogResult.Yes)
+            command = new SqlCommand("Select * FROM Students WHERE ProgramStudent = @cp", mydbCon);
+            command.Parameters.AddWithValue("cp", comboBoxDeleteSelection.Text);
+
+            int exists = Convert.ToInt32(command.ExecuteScalar());
+
+            mydbCon.Close();
+            if (exists > 0)
             {
-                mydbCon = new SqlConnection(@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=
-            C:\Users\TQY_06\Documents\ProjetFinal_CS_BourassaMarie\ProjetFinal_CS_BourassaMarie\FinalProjDB.mdf;
-            Integrated Security=True");
-                mydbCon.Open();
-
-                command = new SqlCommand("DELETE FROM Programs WHERE CodeProgram = @cp", mydbCon);
-                command.Parameters.AddWithValue("cp", comboBoxDeleteSelection.Text);
-                command.ExecuteNonQuery();
-                mydbCon.Close();
+                MessageBox.Show("A student is registered in this program. Please remove the student and try again");
             }
-            MessageBox.Show("Deleted " + comboBoxDeleteSelection.Text + " Successfully");
-            FillCombo();
-           
+            else
+            {
+
+                DialogResult result = MessageBox.Show("Do you really want to delete " + comboBoxDeleteSelection.Text + " ?", "Attention!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+                if (result == DialogResult.Yes)
+                {
+
+                    mydbCon.Open();
+
+                    command = new SqlCommand("DELETE FROM Programs WHERE CodeProgram = @cp", mydbCon);
+                    command.Parameters.AddWithValue("cp", comboBoxDeleteSelection.Text);
+                    command.ExecuteNonQuery();
+                    mydbCon.Close();
+                    MessageBox.Show("Deleted " + comboBoxDeleteSelection.Text + " Successfully");
+                }
+               
+                FillCombo();
+            }
         }
 
         private void buttonBack_Click(object sender, EventArgs e)
         {
             this.Hide();
-            SystemDeGestion addNew = new SystemDeGestion();
-            addNew.Show();
         }
     }
 }
